keiba110/requirements.txt
# HTTP clients
requests>=2.32.0
httpx>=0.27.0

# Data processing
pandas>=2.2.0
polars>=0.20.0
duckdb>=0.10.0
pyarrow>=15.0.0

# PDF processing
tabula-py>=2.9.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Web automation
selenium>=4.19.0
chromedriver-binary-auto>=0.3.0

# Machine learning
lightgbm>=4.3.0
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.44.0

# Dashboard
streamlit>=1.32.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
